fix: install the dependencies and modules you were working with in create-react-app
