+++ package.json
@@ -20,10 +20,9 @@
   },
   "private": false,
   "scripts": {
-    "start": "react-scripts start",
-    "build": "cross-env BABEL_ENV=production babel src -d dist",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "start": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
     "auto": "auto",
     "prepare": "husky install",
     "storybook": "start-storybook -p 6006 -s public",
@@ -61,6 +60,7 @@
     "@storybook/react": "^6.5.14",
     "@storybook/test-runner": "^0.9.1",
     "@storybook/testing-library": "^0.0.13",
+    "@vitejs/plugin-react": "4.0.0",
     "auto": "^10.37.6",
     "babel-plugin-named-exports-order": "^0.0.2",
     "chromatic": "^6.12.0",
@@ -70,6 +70,7 @@
     "lint-staged": "^13.1.0",
     "prettier": "2.8.0",
     "prop-types": "^15.8.1",
+    "vite": "4.3.4",
     "webpack": "^5.75.0"
   },
   "lint-staged": {
