+++ src/components/data/Avatar/stories/Avatar.Default.stories.jsx
