+++ src/components/data/Text/stories/Text.Paragraphs.stories.jsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { Text } from "..";
 
 import {
