+++ src/components/inputs/TextField/stories/TextField.Disabled.stories.jsx
@@ -1,6 +1,7 @@
-import { TextField } from "..";
+import React from "react";
+
+import { TextField, sizes } from "..";
 import { TextFieldController } from "./TextfieldController";
-import { sizes } from "..";
 import { Stack } from "../../../layouts/Stack";
 
 import {
@@ -11,23 +12,16 @@ import {
   placeholder,
   isDisabled,
   type,
-  state,
   value,
   handleChange,
   iconBefore,
   iconAfter,
   maxLength,
   minLength,
-  max,
-  min,
-  isRequired,
-  errorMessage,
-  validMessage,
-  isFullWidth,
 } from "./props";
 
 const story = {
-  title: "inputs/TextField/Size",
+  title: "inputs/TextField/Disabled",
   components: [TextField],
   parameters,
 };
@@ -36,51 +30,40 @@ const TextFieldComponent = (args) => {
   return (
     <Stack justifyContent="space-evenly">
       {sizes.map((size) => (
-        <TextFieldController {...args} key={size} size={size} />
+        <div key={size}>
+          <TextFieldController {...args} size={size} isDisabled={true} />
+        </div>
       ))}
     </Stack>
   );
 };
 
-const Size = TextFieldComponent.bind({});
-Size.args = {
+const Disabled = TextFieldComponent.bind({});
+Disabled.args = {
   label: "Username",
   name: "Username",
   id: "Username",
-  isDisabled: false,
-  placeholder: "Write your full name",
   value: "",
-  state: "pending",
+  placeholder: "Write your full name",
   maxLength: 10,
   minLength: 1,
-  errorMessage: "Please enter only letters in this field",
-  validMessage: "The field has been successfully validated",
-  isFullWidth: false,
-  isRequired: false,
 };
 
-Size.argTypes = {
+Disabled.argTypes = {
   label,
   name,
   id,
   placeholder,
   isDisabled,
   type,
-  state,
   value,
   handleChange,
   iconBefore,
   iconAfter,
   maxLength,
   minLength,
-  max,
-  min,
-  isRequired,
-  errorMessage,
-  validMessage,
-  isFullWidth,
 };
 
 export default story;
 
-export { Size };
+export { Disabled };
