+++ src/components/inputs/TextField/stories/TextField.Invalid.stories.jsx
@@ -1,3 +1,5 @@
+import React from "react";
+
 import { TextField } from "..";
 import { TextFieldController } from "./TextfieldController";
 
@@ -7,55 +9,63 @@ import {
   name,
   id,
   placeholder,
+  isDisabled,
+  state,
+  type,
   value,
   handleChange,
   iconBefore,
   iconAfter,
   maxLength,
   minLength,
-  max,
-  min,
+  errorMessage,
   size,
+  isFullWidth,
 } from "./props";
 
 const story = {
-  title: "inputs/TextField/Default",
+  title: "inputs/TextField/Invalid",
   components: [TextField],
   parameters,
 };
 
-const Default = (args) => <TextFieldController {...args} />;
-Default.args = {
+const Invalid = (args) => <TextFieldController {...args} state="invalid" />;
+Invalid.args = {
   label: "Username",
   name: "Username",
   id: "Username",
-  placeholder: "Write your full name",
-  value: "",
-  maxLength: 10,
+  value: "L-GARZON",
+  placeholder: "Username..",
+  isDisabled: false,
+  type: "text",
+  maxLength: 20,
   minLength: 1,
-  min: 0,
-  max: 0,
+  isRequired: true,
   errorMessage: "Please enter only letters in this field",
   validMessage: "The field has been successfully validated",
   size: "wide",
+  isFullWidth: false,
 };
 
-Default.argTypes = {
+Invalid.argTypes = {
   label,
   name,
   id,
   placeholder,
+  isDisabled,
+  state,
+  type,
   value,
   handleChange,
   iconBefore,
   iconAfter,
   maxLength,
   minLength,
-  max,
-  min,
+  errorMessage,
   size,
+  isFullWidth,
 };
 
 export default story;
 
-export { Default };
+export { Invalid };
