+++ src/components/inputs/TextField/stories/TextField.Number.stories.jsx
@@ -1,70 +1,73 @@
-import { MdSearch } from "react-icons/md";
+import React from "react";
 
 import { TextField } from "..";
 import { TextFieldController } from "./TextfieldController";
 
 import {
   parameters,
+  label,
   name,
   id,
   placeholder,
   isDisabled,
   type,
+  state,
+  value,
   handleChange,
   iconBefore,
   iconAfter,
-  maxLength,
-  minLength,
   max,
   min,
   isRequired,
+  errorMessage,
+  validMessage,
   size,
   isFullWidth,
 } from "./props";
 
 const story = {
-  title: "inputs/TextField/Search",
+  title: "inputs/TextField/Number",
   components: [TextField],
   parameters,
 };
 
-const Search = (args) => (
-  <TextFieldController {...args} type="search" state="pending" />
-);
-Search.args = {
-  name: "searchField",
-  id: "searchField",
-  placeholder: "Search...",
+const Number = (args) => <TextFieldController {...args} />;
+Number.args = {
+  label: "Quantity",
+  name: "quantity",
+  id: "quantity",
+  placeholder: "Value",
+  type: "number",
+  state: "pending",
   isDisabled: false,
-  iconAfter: <MdSearch />,
-  isRequired: false,
-  errorMessage: "",
-  maxLength: 10,
-  minLength: 1,
   max: 10,
-  min: 1,
+  min: 0,
+  isRequired: false,
   size: "wide",
   isFullWidth: false,
 };
 
-Search.argTypes = {
+Number.argTypes = {
+  label,
   name,
   id,
   placeholder,
   isDisabled,
   type,
+  state,
+  value,
   handleChange,
   iconBefore,
   iconAfter,
-  maxLength,
-  minLength,
   max,
   min,
   isRequired,
+  errorMessage,
+  validMessage,
   size,
   isFullWidth,
 };
 
 export default story;
 
-export { Search };
+export { Number };
