+++ src/components/inputs/TextField/stories/TextField.Sizes.stories.jsx
@@ -1,5 +1,8 @@
-import { TextField, sizes } from "..";
+import React from "react";
+
+import { TextField } from "..";
 import { TextFieldController } from "./TextfieldController";
+import { sizes } from "..";
 import { Stack } from "../../../layouts/Stack";
 
 import {
@@ -10,16 +13,23 @@ import {
   placeholder,
   isDisabled,
   type,
+  state,
   value,
   handleChange,
   iconBefore,
   iconAfter,
   maxLength,
   minLength,
+  max,
+  min,
+  isRequired,
+  errorMessage,
+  validMessage,
+  isFullWidth,
 } from "./props";
 
 const story = {
-  title: "inputs/TextField/Disabled",
+  title: "inputs/TextField/Size",
   components: [TextField],
   parameters,
 };
@@ -28,40 +38,51 @@ const TextFieldComponent = (args) => {
   return (
     <Stack justifyContent="space-evenly">
       {sizes.map((size) => (
-        <div key={size}>
-          <TextFieldController {...args} size={size} isDisabled={true} />
-        </div>
+        <TextFieldController {...args} key={size} size={size} />
       ))}
     </Stack>
   );
 };
 
-const Disabled = TextFieldComponent.bind({});
-Disabled.args = {
+const Size = TextFieldComponent.bind({});
+Size.args = {
   label: "Username",
   name: "Username",
   id: "Username",
-  value: "",
+  isDisabled: false,
   placeholder: "Write your full name",
+  value: "",
+  state: "pending",
   maxLength: 10,
   minLength: 1,
+  errorMessage: "Please enter only letters in this field",
+  validMessage: "The field has been successfully validated",
+  isFullWidth: false,
+  isRequired: false,
 };
 
-Disabled.argTypes = {
+Size.argTypes = {
   label,
   name,
   id,
   placeholder,
   isDisabled,
   type,
+  state,
   value,
   handleChange,
   iconBefore,
   iconAfter,
   maxLength,
   minLength,
+  max,
+  min,
+  isRequired,
+  errorMessage,
+  validMessage,
+  isFullWidth,
 };
 
 export default story;
 
-export { Disabled };
+export { Size };
