+++ src/hooks/stories/Example.stories.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Switch } from "../../components/inputs/Switch";
-import { SwitchController } from "./../../components/inputs/Switch/stories/SwitchController";
+import { SwitchController } from "../../components/inputs/Switch/stories/SwitchController";
 import { useMediaQuery } from "../useMediaQuery";
 
 const story = {
