+++ src/shared/typography/stories/Typography.stories.jsx
@@ -1,3 +1,5 @@
+import React from "react";
+
 import { RoleList } from ".";
 
 const story = {
